+++ tests/converters/test_std_msgs.py
@@ -7,7 +7,7 @@ from hypothesis_ros.converters.std_msgs import map_header
 try:
     from std_msgs.msg import Header
 except ImportError:
-    raise ImportError('Please install ROS package ros-{DISTRO}-common-msgs.')
+    raise ImportError('Please install ROS1 package ros-{DISTRO}-common-msgs.')
 import pytest