+++ docs/source/conf.py
@@ -26,7 +26,7 @@ author = u'Florian Kromer'
 # The short X.Y version
 version = u''
 # The full version, including alpha/beta/rc tags
-release = u'0.3.0'
+release = u'1.0.0'
 
 
 # -- General configuration ---------------------------------------------------